[lake] fix checkstyle
