+++ fluss-client/src/test/java/org/apache/fluss/client/table/FlussLakeTableITCase.java
@@ -199,7 +199,9 @@ class FlussLakeTableITCase {
             List<String> lookUpColumns =
                     isDefaultBucketKey
                             ? Arrays.asList("a", "c", "d")
-                            : isPartitioned ? Arrays.asList("a", "d") : Collections.singletonList("a");
+                            : isPartitioned
+                                    ? Arrays.asList("a", "d")
+                                    : Collections.singletonList("a");
             Lookuper lookuper = table.newLookup().lookupBy(lookUpColumns).createLookuper();
             List<InternalRow.FieldGetter> lookUpFieldGetter = new ArrayList<>(lookUpColumns.size());
             for (int columnIndex : pkTableSchema.getColumnIndexes(lookUpColumns)) {
